QuantumPropagators

Documentation for QuantumPropagators.

QuantumPropagators.ChebyWrk

QuantumPropagators.NewtonWrk

QuantumPropagators.arnoldi!

QuantumPropagators.cheby!

QuantumPropagators.cheby_coeffs

QuantumPropagators.cheby_coeffs!

QuantumPropagators.diagonalize_hessenberg_matrix

QuantumPropagators.extend_leja!

QuantumPropagators.extend_newton_coeffs!

QuantumPropagators.initpropwrk

QuantumPropagators.newton!

QuantumPropagators.propagate

QuantumPropagators.propstep!

QuantumPropagators.ChebyWrk — Type

Workspace for the Chebychev propagation routine.

    ChebyWrk(Ψ, Δ, E_min, dt; limit=1e-12)

initializes the workspace for the propagation of a state similar to Ψ under a Hamiltonian with eigenvalues between E_min and E_min + Δ, and a time step dt. Chebychev coefficients smaller than the given limit are discarded.

QuantumPropagators.NewtonWrk — Type

    NewtonWrk(v0, m_max=10)

Workspace for the Newton-with-restarted-Arnoldi propagation routine.

Initializes the workspace for the propagation of a vector v0, using a maximum Krylov dimension of m_max in each restart iteration. Note that m_max should be smaller than the length of v0.

QuantumPropagators.arnoldi! — Method

arnoldi!(Hess, q, m, Ψ, H, dt=1.0; extended=true, norm_min=1e-15)

Calculate the Hessenberg matrix and Arnoldi vectors of H dt, from Ψ.

For a given order m, the m×m Hessemberg matrix is calculated and stored in in the pre-allocated Hess. Further an array of m normalized Arnoldi vectors is stored in in the pre-allocated q, plus one additional unnormalized Arnoldi vector. The unnormalized m+1st vector could be used to easily extend a given m×m Hessenberg matrix to a (m+1)×(m+1) matrix.

If the extended Hessenberg matrix is requested (extended=true, default), the m+1st Arnoldi vector is also normalized, and it's norm will be stored in m+1, m entry of the (extended) Hessenberg matrix, which is an (m+1)×(m+1) matrix.

Return the size m of the calculated Hessenberg matrix. This will usually be the input m, except when the Krylov dimension of H starting from Ψ is less then m. E.g., if Ψ is an eigenstate of H, the returned m will be

See http://en.wikipedia.org/wiki/Arnoldi_iteration for a description of the algorithm.

Arguments

Hess::Matrix{ComplexF64}: Pre-allocated storage for the Hessemberg matrix. Can be uninitialized on input. The matrix must be at least of size m×m, or (m+1)×(m+1) if extended=true. On output, the m×m sub-matrix of Hess (with the returned output m) will contain the Hessenberg matrix, and all other elements of Hess be be set to zero.

q: Pre-allocated array of states similar to Ψ, as storage for the calculated Arnoldi vectors. These may be un-initialized on input. Must be at least of length m+1

m: The requested dimensions of the output Hessenberg matrix.

Ψ: The starting vector for the Arnoldi procedure. This can be of any type, as long as Φ = H * Ψ results in a vector similar to Ψ, there is an inner products of Φ and Ψ (Ψ⋅Φ is defined), and norm(Ψ) is defined.

H: The operator (up to dt) for which to calculate the Arnoldi procedure. Can be of any type, as long as H * Ψ is defined.

dt: The implicit time step; the total operator for which to calculate the Arnoldi procedure is H * dt

extended: If true (default), calculate the extended Hessenberg matrix, and normalized the final Arnoldi vector

norm_min: the minimum value of the norm of Ψ at which Ψ should be considered the zero vector

QuantumPropagators.cheby! — Method

Evaluate Ψ = exp(-i H dt) Ψ in-place.

Args:

Ψ: on input, initial vector. Will be overwritten with result.

H: Hermitian operator

dt: time step

E_min: minimum eigenvalue of H, to be used instead of the E_min from the initialization of wrk. The same wrk may be used for different values E_min, as long as the spectra radius Δ and the time step dt are the same as those used for the initialization of wrk.

The routine will not allocate any internal storage. This implementation requires copyto! lmul!, and axpy! to be implemented for Ψ, and the three-argument mul! for Ψ and H.

QuantumPropagators.cheby_coeffs! — Method

Calculate Chebychev coefficients in-place.

Writes the coefficients into coeffs. Returns the number n of coefficients required for convergence, or the size of coeffs, whichever is less.

coeffs: An array to hold the Chebychev cofficients. On output, the elements 1 through n will hold the calculated coefficients, while the remaining elements will be unchanged

dt: the time step

QuantumPropagators.cheby_coeffs — Method

Calculate Chebychev coefficients.

Return an array of coefficiencts larger than limit.

Args:

Δ: the spectral radius of the underlying operator

dt: the time step

QuantumPropagators.diagonalize_hessenberg_matrix — Method

diagonalize_hessenberg_matrix(Hess, m; accumulate=false)

Diagonalize the m × m top left submatrix of the given Hessenberg matrix.

If accumulate is true, return the concatenated eigenvalues for Hess[1:1,1:1] to Hess[1:m,1:m], that is, all sumatrices of size 1 through m.

QuantumPropagators.extend_leja! — Method

extend_leja!(leja, n, newpoints, n_use)

Given an array of n (ordered) Leja points, extract n_use points from newpoints, and append them to the existing Leja points. The array leja should be sufficiently large to hold the new Leja points, which are appended after index n_old. It will be re-allocated if necessary and may have a size of up to 2*(n+n_use).

Arguments

leja: Array of leja values. Must contain the "old" leja values to be kept in leja(0:n-1). On output, n_use new leja points will be in leja(n+:n+n_use-1), for the original value of n. The leja array must use zero-based indexing.

n: On input, number of "old" leja points in leja. On output, total number of leja points (i.e. n=n+n_use)

newpoints: On input, candidate points for new leja points. The n_use best values will be chosen and added to leja. On output, the values of new_points are undefined.

n_use: Number of points that should be added to leja

QuantumPropagators.extend_newton_coeffs! — Method

extend_newton_coeffs!(a, n_a, leja, func, n_leja, radius)

Extend the array a of existing Newton coefficients for the expansion of the func from n_a coefficients to n_leja coefficients. Return a new value n_a=n_a+n_leja with the total number of Newton coefficients in the updated a.

Arguments

a: On input, a zero-based array of length n_a or greater, containing Newton coefficients. On output, array containing a total n_leja coefficients. The array a will be resized if necessary, and may have a length greater than n_leja on output

n_a: The number of Newton coefficients in a, on input. Elements of a beyond the first n_a elements will be overwritten.

leja: Array of normalized Leja points, containing at least n_leja elements.

func: Function for which to calcluate Newton coeffiecients

n_leja: The number of elements in leja to use for calculating new coefficients, and the total number of Newton coefficients on output

radius: Normalization radius for divided differences

QuantumPropagators.initpropwrk — Method

Initialize a workspace for propagation.

    wrk = initpropwrk(state, tlist, generator...; method=:auto, kwargs...)

The resulting wrk can be passed to propagate or propstep.

Arguments

state: An exemplary state for the propagation (e.g., the initial state)

tlist: The time grid over which propagate will be called

generator: An exemplary dynamical generator. For full generality (if method=:cheby, the given generator should have a spectral range sufficiently large to encompass the entire propagation. If given multiple times, a spectral envelope enclosing all the generators will be determined automatically. In this case, you should pass the generators with the extremal values of all the controls.

method: The propagation method to use. The default value of :auto attempts to choose the best method available, based on the properties of the given state, tlist, and generator. Alternative values are :cheby and :newton.

QuantumPropagators.newton! — Function

newton!(Ψ, H, dt, wrk, func=(z -> exp(-1im*z)); norm_min=1e-14, relerr=1e-12,
        max_restarts=50)

Evaluate Ψ = func(H*dt) Ψ using a Newton-with-restarted-Arnoldi scheme.

Arguments

Ψ: The state to propagate, will be overwritten in-place with the propagated state

H: Operator acting on Ψ. Together with dt, this is the argument to func

dt: Implicit time step. Together with H, this is the argument to func

wkr: Work array, initialized with NewtonWrk

func: The function to apply to H dt, taking a single (scalar) complex-valued argument z in place of H dt. The default func is to evaluate the time evoluation operator for the Schrödinger equation

norm_min: the minium norm at which to consider a state similar to Ψ as zero

relerr: The relative error defining the convergence condition for the restart iteration. Propagation stops when the norm of the accumulated Ψ is stable up to the given relative error

max_restart: The maximum number of restart iterations. Exceeding max_restart will throw an AssertionError.

QuantumPropagators.propagate — Method

Propagator a state over an entire time grid

    propagate(state0, genfunc, tlist; wrk=nothing, storage=nothing)

Propagates state0 over the tlist time grid, using genfunc(t) as the dynamical generator with t being the midpoint of each time interval. Return the propagated state at tlist[end].

The propagation method is determined by wrk, see initipropwrk. If wrk is not given, propagate will attempt to choose the most appropriate method for the given parameters.

If storage is given, the propagated states at each point in tlist will be written to it.

QuantumPropagators.propstep! — Method

Perform a single propagation step.

    propstep!(state, generator, dt, wrk;, kwargs...)

The propagation method is determined by wrk, see initpropwrk. The kwargs are forwarded to the underlying method